% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_prop.R
\name{attach_xyDist}
\alias{attach_xyDist}
\title{Attaches y and x distance from radar in km to object location indices}
\usage{
attach_xyDist(obj_props, xdist, ydist)
}
\description{
Attaches y and x distance from radar in km to object location indices
}
