% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_obj.R
\name{attach_xyheads}
\alias{attach_xyheads}
\title{saves last x y movements of the objects.}
\usage{
attach_xyheads(frame1, frame2, current_objects)
}
\description{
Attaches last xyheads to current objects for future use.
}
