% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity.R
\name{check_bigOverlap}
\alias{check_bigOverlap}
\title{checks overlapping regoin for big objects in both the frames.}
\usage{
check_bigOverlap(obj_extend, target_extend)
}
\description{
Checks overlapping area in pixels, size of the object and return if overlapping is considerable.
}
