% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_split.R
\name{check_merging}
\alias{check_merging}
\title{Checks possible merging of the dead objects.}
\usage{
check_merging(dead_obj_id1, current_objects, frame1, frame2)
}
\description{
This function takes in two R-lists containing information about current objects
in the frame1 and their properties, such as center location and area. If the
I am using an arbitrary crieterion for merging. If euclidean distance between centers
of the two objects c_dist < or = to r=sqrt(area), then merging is considered.
Here, if we assume square objects, then the r is length of a sides of the square.
If all objects are dead in frame2 then no merging happened.
}
