% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_prop.R
\name{check_searchBox}
\alias{check_searchBox}
\title{checks that the search box is in the domain.}
\usage{
check_searchBox(search_box, img_dims)
}
\description{
Returns NA if search box  outside the image or search box is very small.
}
