% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_preproc.R
\name{clear_smallEchoes}
\alias{clear_smallEchoes}
\title{Removed objects smaller than the given size.}
\usage{
clear_smallEchoes(label_image, min_size)
}
\description{
Takes in labeled image removes objects smaller than min_size and returns re-labeled image.
Pixels attached diagonally are not considered as continuouse part of the object.
}
