% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_prop.R
\name{correct_shift}
\alias{correct_shift}
\title{FFT shifts are corrected using last headings.}
\usage{
correct_shift(this_shift, current_objects, object_id1)
}
\description{
takes in flow vector based shift and current_object dataframe which has last
headings, and check if they are resonably close if not rejects or modify shift and return.
Note:  frame2 of last timestep is now frame1, but current_objects still has it as frame2.
So id2 in the last frame2 are actually ids related to frame1 now.
}
