% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outNC.R
\name{create_outNC_track}
\alias{create_outNC_track}
\title{Creates output netcdf file for radar echo trajectories.}
\usage{
create_outNC_track(ofile, max_obs)
}
\arguments{
\item{ofile}{name and path of the output file.}

\item{max_obs}{longest possible track record. Deafult maximum 65 observation per track.}
}
\description{
This is the longest function.
}
