% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded_clones.R
\name{expanded_clones}
\alias{expanded_clones}
\title{Identifying Expanded Clones}
\usage{
expanded_clones(
  seurat_object,
  clonotype_column,
  sample_column,
  sample.1,
  sample.2,
  minimum_count = 0,
  decreasing = TRUE,
  identity = NULL
)
}
\arguments{
\item{seurat_object}{seurat object to operate on}

\item{clonotype_column}{Character describing name of column to use for clone calling}

\item{sample_column}{Character describing name of column to use for sample discrimination}

\item{sample.1}{Character describing a sample in 'sample_column' for comparison}

\item{sample.2}{Character describing a sample in 'sample_column' for comparison}

\item{minimum_count}{integer value specifying the minimum number of clones for a clonotype to be included in analysis}

\item{decreasing}{Boolean value specifying whether or not to sort by decreasing (or increasing) expansion}

\item{identity}{A value in 'clonotype_identity_mode' to specify clonotypes of interest (must first run the 'calculate_freq' function)}
}
\value{
a dataframe of clones sorted by their expansion in sample.1 compared to sample.2
}
\description{
This function takes a seurat object with meta data that already includes
TCR annotations and calculates the expansion of clones in sample.1 compared to sample.2
}
