% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_flow.R
\name{fft_crossCov}
\alias{fft_crossCov}
\title{Computes cross-covariance using FFT method, returns shifted covariance image}
\usage{
fft_crossCov(img1, img2)
}
\description{
Computes cross-covariance using FFT method, returns shifted covariance image
}
