% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_flow.R
\name{fft_flowVectors}
\alias{fft_flowVectors}
\title{Estimates flow vectors in two images using cross covariance of the images.}
\usage{
fft_flowVectors(im1, im2)
}
\description{
Leese, John A., Charles S. Novak, and Bruce B. Clark.
          "An automated technique for obtaining cloud motion from geosynchronous
          satellite data using cross correlation."
          Journal of applied meteorology 10.1 (1971): 118-132.
}
