% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_flow.R
\name{fft_shift}
\alias{fft_shift}
\title{Rearranges the crossCov matrix}
\usage{
fft_shift(fft_mat)
}
\description{
Rearranges the crossCov matrix so that 'zero' frequency or DC component
 is in the middle of the matrix. Taken from stackoverflow Que. 30630632
}
