% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_preproc.R
\name{filterFrame}
\alias{filterFrame}
\title{filters frame for small objects when frame is already available.}
\usage{
filterFrame(frame, min_size = 2)
}
\description{
filters frame for small objects when frame is already available.
}
