% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_prop.R.R
\name{find_objects}
\alias{find_objects}
\title{Returns vector of objects ids in the given reion.}
\usage{
find_objects(search_box, image2)
}
\description{
Given the search box and image2, returns objects in the region.
}
