% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_prop.R
\name{fitEllipse}
\alias{fitEllipse}
\title{Try to fit optimum ellipse for the object circularity, given the object index.
The ellipse fitted with this method does not enclosed the object but it provide optimum ellipse parameters.
The ratio of major and minor axis and eccentricity are correctly estimated for tracking purpose.}
\usage{
fitEllipse(object_index)
}
\description{
Try to fit optimum ellipse for the object circularity, given the object index.
The ellipse fitted with this method does not enclosed the object but it provide optimum ellipse parameters.
The ratio of major and minor axis and eccentricity are correctly estimated for tracking purpose.
}
