% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity.R
\name{get_disparity}
\alias{get_disparity}
\title{Actually computes desparity for a single object.}
\usage{
get_disparity(obj_found, image2, search_box, obj1_extent)
}
\description{
Check how it is computed for detail.
This parameter has most effect on the acccuracy of the tracks.
}
