% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity.R
\name{get_disparity_all}
\alias{get_disparity_all}
\title{Returns Disparity of all the objects in the region.}
\usage{
get_disparity_all(obj_found, image2, search_box, obj1_extent)
}
\description{
Returns disparities of all the objects found within the search box or NA if
no object is present.
}
