% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_preproc.R
\name{get_filteredFrame}
\alias{get_filteredFrame}
\title{Returns a single radar scan from the input netcdf file.}
\usage{
get_filteredFrame(ncfile, var_id, scan_num, min_size = 2)
}
\arguments{
\item{ncfile}{input netcdf file object from ncdf4 package.}

\item{var_id}{string name of the varibale in the file.}

\item{scan_num}{index of frame to be read from the file.}

\item{min_size}{in pixels objects smaller than this will be removed. Default 2.}
}
\description{
Smaller objects are removed and the rest are lebeled.
}
