% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{get_matchPairs}
\alias{get_matchPairs}
\title{returns pairs of object ids from both frames.}
\usage{
get_matchPairs(image1, image2)
}
\description{
Given two images, the function identifies the matching 
objects and pair them appropriatly using Hungarian method and desparity.
}
\seealso{
code{disparity function}.
}
