% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_flow.R
\name{get_objAmbientFlow}
\alias{get_objAmbientFlow}
\title{Computes flow in the vicinity of the object.}
\usage{
get_objAmbientFlow(obj_extent, img1, img2, margin)
}
\description{
Takes in object info (major axis and center) and two images to estimate ambient flow using FFT phase correlation.
Margin is the additional region arround the object used to comput the flow vectors.
}
