% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_prop.R
\name{get_objectCenter}
\alias{get_objectCenter}
\title{return center indices of the object.}
\usage{
get_objectCenter(obj_id, labeled_image)
}
\description{
Returns indices of center pixel of the given object id from a labeled image.
This may be done in better way for non-oval objects.
}
