% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_prop.R.R
\name{get_objectProp}
\alias{get_objectProp}
\title{Return all the object's size, location and classification info.}
\usage{
get_objectProp(image1, xyDist)
}
\description{
xyDist should be a list of x_dist and y_dist in km.
}
