% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_prop.R
\name{get_objExtent}
\alias{get_objExtent}
\title{Returns object extent properties.}
\usage{
get_objExtent(labeled_image, obj_label)
}
\description{
Takes in a labeled image and finds the radius, area and the center of the given object.
}
