% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_split.R
\name{get_origin_uid}
\alias{get_origin_uid}
\title{Find id of the parent of the new born object.}
\usage{
get_origin_uid(obj, frame1, old_objects)
}
\description{
returns unique id of the origin (or zero) for given object in frame1.
Also remember that old object id2 is actual id1 in frame1, as we still have
to update the object_ids.
}
