% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity.R
\name{get_sizeChange}
\alias{get_sizeChange}
\title{returns size change between the frames.}
\usage{
get_sizeChange(x, y)
}
\description{
Returns change in size of the eacho as ratio of bigger number to smaller
number when given two number, minus 1.
}
