% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_flow.R
\name{get_std_flowVector}
\alias{get_std_flowVector}
\title{Alternative to get_objAmbientFlow.}
\usage{
get_std_flowVector(obj_extent, img1, img2, margin, magnitude)
}
\description{
Flow vectors magnitude is clipped to given magnitude to controll erratic output from FFT flow.
}
