% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_obj.R
\name{init_uids}
\alias{init_uids}
\title{Returns a dataframe for objects with unique ids and their corresponding ids in frame1 and frame2.}
\usage{
init_uids(first_frame, second_frame, pairs)
}
\arguments{
\item{first_frame}{First image for tracking.}

\item{second_frame}{Second image for tracking.}

\item{pairs}{output of \code{get_match_pairs()}}
}
\description{
This function is called when new rainy scan is seen after the period of no rain or the first time.
}
