% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{locate_allObjects}
\alias{locate_allObjects}
\title{Matches all the obejects in image1 to the objects in image2.}
\usage{
locate_allObjects(image1, image2)
}
\description{
This is the main function to be called on two sets of radar images, for tracking.
}
