% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{match_pairs}
\alias{match_pairs}
\title{Matches objects into pairs and removes bad matching.}
\usage{
match_pairs(obj_match)
}
\description{
The bad matching is when disparity is more than the set value.
}
