% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_obj.R
\name{next_uid}
\alias{next_uid}
\title{Returns sequence of next unique ids and increament the uid_counter.}
\usage{
next_uid(count = 1)
}
\description{
Returns sequence of next unique ids and increament the uid_counter.
}
