% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_objects_label}
\alias{plot_objects_label}
\title{Plots image with objects labels.}
\usage{
plot_objects_label(labeled_image, xvalues, yvalues)
}
\description{
This is used in development stage to test images when processed 1-by-1.
}
