% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_prop.R
\name{predict_searchExtent}
\alias{predict_searchExtent}
\title{predict search region.}
\usage{
predict_searchExtent(obj1_extent, shift, search_radius)
}
\description{
Predicts search extent/region for the object in image2 given the image shift.
}
