% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{save_objMatch}
\alias{save_objMatch}
\title{Corrects and saves disparity for the object matching stage.}
\usage{
save_objMatch(obj_id1, obj_found, disparity, obj_match)
}
\description{
If disparity is large then it saves a large number for the value to reduce
the chances of this pairing to zero, else it save the value in the obj_match array.
}
