% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_obj.R
\name{survival_stats}
\alias{survival_stats}
\title{Returns a list with number of objects lived, died and born between the current and the previousstep.}
\usage{
survival_stats(pairs, num_obj2)
}
\description{
Returns a list with number of objects lived, died and born between the current and the previousstep.
}
