% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outNC.R
\name{write_duration}
\alias{write_duration}
\title{Write duration of dead objects in output NC file.}
\usage{
write_duration(outNC, current_objects)
}
\description{
Writes number of observations for dead objects. Duration is in time-steps.
}
