% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outNC.R
\name{write_settingParms_toNC}
\alias{write_settingParms_toNC}
\title{Writes all the setting parameters (as attributes) for the tracking.}
\usage{
write_settingParms_toNC(outNC)
}
\description{
These parameters affect the sensitivity of the tracks, mergers and split definitions etc.
}
