% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outNC.R
\name{write_survival}
\alias{write_survival}
\title{Write survival stats}
\usage{
write_survival(outNC, survival_stat, time, scan)
}
\description{
write number of lived, dead and born objects to the file for each scan.
}
