% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outNC.R
\name{write_update}
\alias{write_update}
\title{Writes properties and uids for all objects into output netcdf file.}
\usage{
write_update(outNC, current_objects, obj_props, obs_time)
}
\arguments{
\item{outNC}{output netcdf file object from function \code{\link{create_outNC_track}}}

\item{current_objects}{output of \code{update_current_objects}}

\item{obj_props}{output of \code{get_object_prop()}}

\item{obs_time}{time of first scan in POSIX format. units="seconds since 1970-01-01".}
}
\description{
Writes properties and uids for all objects into output netcdf file.
}
